requirements.txt
streamlit>=1.10.0
altair>=4.0.0,<5
pandas>=1.3.0
google-genai>=1.0.0
python-dotenv>=1.0.0
plotly>=5.0.0
pytest>=7.0.0
protobuf==3.20.*